requirements.txt
numpy==1.24.4
openpyxl==3.1.2
pandas==2.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tzdata==2023.3
python_docx==1.1.0
